+++ host/channels/gmail_channel.py
@@ -13,7 +13,7 @@ from google_auth_oauthlib.flow import InstalledAppFlow
 from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
 
-from . import register_channel
+from . import register_channel_class as register_channel
 from .. import config
 from ..env import read_env_file