+++ host/channels/slack_channel.py
@@ -5,7 +5,7 @@ from typing import Callable, Optional
 from slack_bolt.async_app import AsyncApp
 from slack_bolt.adapter.socket_mode.async_handler import AsyncSocketModeHandler
 
-from . import register_channel
+from . import register_channel_class as register_channel
 from .. import config
 from ..env import read_env_file